+++ mergeSort.cpp
@@ -4,14 +4,14 @@
 
 using namespace std;
 
-pair<vector<int>, pair<int, int>> interleave(vector<int> A, int initial_index, int middle_index, int end_index, int counter_comparisons, int counter_movements);
-pair<vector<int>, pair<int, int>> mergeSort(vector<int> array);
+pair<vector<long long>, pair<long long, long long>> interleave(vector<long long> A, long long initial_index, long long middle_index, long long end_index, long long counter_comparisons, long long counter_movements);
+pair<vector<long long>, pair<long long, long long>> mergeSort(vector<long long> array);
 
-pair<vector<int>, pair<int, int>> interleave(vector<int> A, int initial_index, int middle_index, int end_index, int counter_comparisons, int counter_movements) {
-    vector<int> B(end_index - initial_index + 1);
+pair<vector<long long>, pair<long long, long long>> interleave(vector<long long> A, long long initial_index, long long middle_index, long long end_index, long long counter_comparisons, long long counter_movements) {
+    vector<long long> B(end_index - initial_index + 1);
 
     // Copy elements from A to B
-    for (int i = initial_index; i <= middle_index; ++i) {
+    for (long long i = initial_index; i <= middle_index; ++i) {
         counter_comparisons++;
 
         counter_movements++;
@@ -19,7 +19,7 @@ pair<vector<int>, pair<int, int>> interleave(vector<int> A, int initial_index, i
     }
 
     // Copy elements from A to B in reverse order
-    for (int j = middle_index + 1; j <= end_index; ++j) {
+    for (long long j = middle_index + 1; j <= end_index; ++j) {
         counter_comparisons++;
 
         counter_movements++;
@@ -27,11 +27,11 @@ pair<vector<int>, pair<int, int>> interleave(vector<int> A, int initial_index, i
     }
 
     counter_movements += 2;
-    int i = initial_index;
-    int j = end_index;
+    long long i = initial_index;
+    long long j = end_index;
 
     // Merge sorted halves from B back into A
-    for (int k = initial_index; k <= end_index; ++k) {
+    for (long long k = initial_index; k <= end_index; ++k) {
         counter_comparisons++;
 
         counter_comparisons++;
@@ -49,25 +49,25 @@ pair<vector<int>, pair<int, int>> interleave(vector<int> A, int initial_index, i
     return make_pair(A, make_pair(counter_comparisons, counter_movements));
 }
 
-pair<vector<int>, pair<int, int>> mergeSort(vector<int> array) {
+pair<vector<long long>, pair<long long, long long>> mergeSort(vector<long long> array) {
 
-    int counter_comparisons = 0;
-    int counter_movements = 0;
+    long long counter_comparisons = 0;
+    long long counter_movements = 0;
 
     counter_movements += 3;
-    vector<int> A = array;
-    int n = array.size();
-    int i = 1;
+    vector<long long> A = array;
+    long long n = array.size();
+    long long i = 1;
 
     while (i < n) {
         counter_comparisons++;
 
-        for (int initial_index = 0; initial_index < n; initial_index += 2 * i) {
+        for (long long initial_index = 0; initial_index < n; initial_index += 2 * i) {
             counter_comparisons++;
 
             counter_movements += 2;
-            int middle_index = initial_index + i - 1;
-            int end_index = min(initial_index + 2 * i - 1, n - 1);
+            long long middle_index = initial_index + i - 1;
+            long long end_index = min(initial_index + 2 * i - 1, n - 1);
 
             counter_comparisons++;
             if (middle_index < end_index) {
