+++ quickSort.cpp
@@ -4,17 +4,17 @@
 
 using namespace std;
 
-pair<vector<int>, pair<int, int>> quickSort(vector<int> array);
-pair<vector<int>, pair<int, int>> quicksort_ordena(vector<int> array, int p, int r, int counter_comparisons, int counter_movements);
-tuple<vector<int>, int, int, pair<int, int>> quicksort_particao_ternaria(vector<int> array, int p, int r, int counter_comparisons, int counter_movements);
+pair<vector<long long>, pair<long long, long long>> quickSort(vector<long long> array);
+pair<vector<long long>, pair<long long, long long>> quicksort_ordena(vector<long long> array, long long p, long long r, long long counter_comparisons, long long counter_movements);
+tuple<vector<long long>, long long, long long, pair<long long, long long>> quicksort_particao_ternaria(vector<long long> array, long long p, long long r, long long counter_comparisons, long long counter_movements);
 
-pair<vector<int>, pair<int, int>> quickSort(vector<int> array) {
+pair<vector<long long>, pair<long long, long long>> quickSort(vector<long long> array) {
 
-    int counter_comparisons = 0;
-    int counter_movements = 0;
+    long long counter_comparisons = 0;
+    long long counter_movements = 0;
 
     counter_movements++;
-    int n = array.size();
+    long long n = array.size();
 
     auto result = quicksort_ordena(array, 0, n - 1, counter_comparisons, counter_movements);
     counter_comparisons = result.second.first;
@@ -24,14 +24,14 @@ pair<vector<int>, pair<int, int>> quickSort(vector<int> array) {
     return make_pair(result.first, make_pair(counter_comparisons, counter_movements));
 }
 
-tuple<vector<int>, int, int, pair<int, int>> quicksort_particao_ternaria(vector<int> array, int p, int r, int counter_comparisons, int counter_movements) {
+tuple<vector<long long>, long long, long long, pair<long long, long long>> quicksort_particao_ternaria(vector<long long> array, long long p, long long r, long long counter_comparisons, long long counter_movements) {
 
     counter_movements += 3;
-    int x = array[r];
-    int i = p - 1;
-    int k = r;
+    long long x = array[r];
+    long long i = p - 1;
+    long long k = r;
 
-    for (int j = p; j < k; ++j) {
+    for (long long j = p; j < k; ++j) {
         counter_comparisons++;
 
         if (array[j] < x) {
@@ -51,9 +51,9 @@ tuple<vector<int>, int, int, pair<int, int>> quicksort_particao_ternaria(vector<
     }
 
     counter_movements++;
-    int q = i + 1;
+    long long q = i + 1;
 
-    for (int j = k; j <= r; ++j) {
+    for (long long j = k; j <= r; ++j) {
         counter_comparisons++;
 
         counter_movements += 2;
@@ -62,20 +62,20 @@ tuple<vector<int>, int, int, pair<int, int>> quicksort_particao_ternaria(vector<
     }
 
     counter_movements++;
-    int t = i;
+    long long t = i;
 
     return make_tuple(array, q, t, make_pair(counter_comparisons, counter_movements));
 }
 
-pair<vector<int>, pair<int, int>> quicksort_ordena(vector<int> array, int p, int r, int counter_comparisons, int counter_movements) {
+pair<vector<long long>, pair<long long, long long>> quicksort_ordena(vector<long long> array, long long p, long long r, long long counter_comparisons, long long counter_movements) {
     while (p < r) {
         counter_comparisons++;
 
         auto result = quicksort_particao_ternaria(array, p, r, counter_comparisons, counter_movements);
 
         array = get<0>(result);
-        int q = get<1>(result);
-        int t = get<2>(result);
+        long long q = get<1>(result);
+        long long t = get<2>(result);
         counter_comparisons = get<3>(result).first;
         counter_movements = get<3>(result).second;
         counter_movements += 3;
