+++ selectionSort.h
@@ -4,6 +4,6 @@
 #include <vector>
 #include <utility>
 
-std::pair<std::vector<int>, std::pair<int, int>> selectionSort(std::vector<int> array);
+std::pair<std::vector<long long>, std::pair<long long, long long>> selectionSort(std::vector<long long> array);
 
 #endif // SELECTION_SORT_H
